Fix loading of font files from bower_components
The relative URIs for font files in the vulcanized font-roboto component
could not be resolved properly at runtime due to the way core elements
are imported (via app:// URIs). To get around this the vulcanized
components bundle is now generated inside the bower_components
directory. Since the bundle is required by all core elements it is now
loaded explicitly in RendererContext so there's no need to HTML import
it in every element.