Use a JavaScript-free workaround for a Chrome flexbox bug
When the page element (which is a flex container) is embedded in
a panel element (also a flex container) the flex children in the page
element don't grow to fill the page. This appears to be a known bug in
Chrome <https://code.google.com/p/chromium/issues/detail?id=542388>.
The workaround involves wrapping the nested flex container in an
absolutely positioned div.