Fix resizing and background color of the directory tree view
Previously the directory tree view in the workspace had a fixed size so
when the page tree element above it was resized by moving the splitter
the directory tree view did not auto-resize to fit the available space.
These changes make the directory tree view take up all the space below
the page tree element.