Fix misnamed options being used to create BrowserWindow(s)
BrowserWindow option names were normalized in v0.35. I forgot to update
this bit after the recent update to the Electron typings, and TypeScript
didn't pick up the errors because strict checking of object literals
only occurs when the variable the object literal is assigned to
(`options` in this case) is declared with an explicit type
(and it wasn't in this case).