Use the `node-pre-gyp` workaround in `electron-rebuild`
The workaround consists of copying the built native modules to the
directory where `node-pre-gyp` looks for them at runtime. This seems
to be the only way to get `node-inspector` working properly because my
attempts to monkey-patch `node-pre-gyp` seem to fail in that case
(even thought monkey-patching worked for other native modules like
`sqlite3`). This workaround can be removed if
https://github.com/mapbox/node-pre-gyp/pull/187 is ever merged in.