Convert the dialog used for editing a debug configuration to a tab page
This is just the first set of changes, there are a couple of major
issues left to address. Changes aren't properly discarded when a debug
config page is closed, so any attempt to reopen a debug config for
editing after it was closed will fail. And, the styling is quite messed
up, but that's mostly due to the lack of a coherent theme.