Discard changes when a debug configuration page is closed
To do this neatly I ended up redoing how page elements work, from now on
there'll be a separate page element class for each type of content that
can be presented on a page, so now there is a code editor page element,
and a debug config editor page element. The base functionality of page
elements has been moved to a behavior that will be reused by all the
various page element classes.