Update react-free-style to v2.2.0
This new release has a "typings" field in its `package.json` so the
TypeScript compiler can locate the typings by itself. The typings that
are shipped with the package have to be adjusted to accept an additional
parameter for the component context, this is done by overwriting the
`.d.ts` in the package with the customized one stored in `app/typings`.