Switch to common React type definitions from DefinitelyTyped
Maintaining a customized version of the React type definitions was a
waste of time since it required customization of the type definitions
of every third-party library that relied on React type definitions.
The only reason the customized version of the type definitions was
created in the first place was to add an additional type parameter to
React.Component so that the context property could be properly typed in
component classes that rely on context without having to redefine the
context property in each of those classes.

From here on component classes that need a properly typed context
property can be derived from ContextComponent that has a type parameter
to specify the context type. This is a pretty lame workaround because
ContextComponent really doesn't do anything, but it'll do for now.