Renderer IPC dispatcher didn't check the nodeId in simple messages
When a message has a nodeId set it's meant to be dispatched only to the
node matching the id, but the renderer dispatcher didn't check the id
and simply dispatched the message to all nodes subscribed to the message
key.

Tests have been modified to catch this particular issue in case it
shows up again.